seperate common value into application.yml, remove unused codes, extract feign setting to create application-search
