+++ blog-api/build.gradle
@@ -19,29 +19,26 @@ repositories {
 }
 
 ext {
-    // set('springCloudVersion', "2021.0.6")
+    set('springCloudVersion', "2021.0.6")
     set('mapstructVersion', "1.4.2.Final")
 }
 
-/*
+
 dependencyManagement {
     imports {
         mavenBom "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}"
     }
 }
-*/
+
 
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-web'
-    // implementation 'org.springframework.cloud:spring-cloud-starter-openfeign'
+    implementation 'org.springframework.cloud:spring-cloud-starter-openfeign'
 
-    implementation 'org.springdoc:springdoc-openapi-ui:1.6.3'
+    implementation 'org.springdoc:springdoc-openapi-ui:1.6.15'
 
     implementation "org.mapstruct:mapstruct:${mapstructVersion}"
     implementation 'org.projectlombok:lombok-mapstruct-binding:0.2.0'
-    // implementation 'com.google.inject:guice:5.0.1'
-
-    implementation 'com.github.ulisesbocchio:jasypt-spring-boot-starter:3.0.4'
 
     implementation 'org.springframework.boot:spring-boot-starter-data-redis'
     implementation group: 'it.ozimov', name: 'embedded-redis', version: '0.7.1'
@@ -56,6 +53,8 @@ dependencies {
     testImplementation 'org.springframework.restdocs:spring-restdocs-mockmvc'
 
     implementation group: 'org.springframework.boot', name: 'spring-boot-starter-aop', version: '2.5.5'
+
+    implementation 'com.github.ulisesbocchio:jasypt-spring-boot-starter:3.0.5'
 
     implementation project(':blog-search')
     implementation project(':blog-common')
