+++ blog-api/src/main/java/bank/common/config/WebConfig.java
@@ -1,7 +1,6 @@
 package bank.common.config;
 
 import bank.controller.handler.SortTypeConverter;
-import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.format.FormatterRegistry;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
