+++ blog-api/src/main/java/bank/controller/v1/mapper/PopularKeywordV1Mapper.java
@@ -2,7 +2,7 @@ package bank.controller.v1.mapper;
 
 import bank.controller.v1.dto.PopularKeywordV1;
 import bank.domain.keyword.PopularKeyword;
-import bank.remote.common.GenericMapper;
+import bank.mapper.GenericMapper;
 import org.mapstruct.Mapper;
 import org.mapstruct.factory.Mappers;
