+++ blog-api/src/main/java/bank/service/keyword/RegisterPopularKeywordServiceImpl.java
@@ -1,6 +1,7 @@
 package bank.service.keyword;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 
@@ -8,12 +9,16 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class RegisterPopularKeywordServiceImpl implements RegisterPopularKeywordService {
 
+    @Value("${spring.redis.key.keyword}")
+    private String key;
+
+
     private final RedisTemplate<String, Object> redisTemplate;
 
     @Override
     public void processIfPresentOrNot(final String query) {
         try {
-            redisTemplate.opsForZSet().incrementScore("keyword", query, 1);
+            redisTemplate.opsForZSet().incrementScore(key, query, 1);
         } catch (Exception e) {
             e.printStackTrace();
         }
