+++ blog-api/src/main/resources/application-h2.yaml
@@ -1,0 +1,3 @@
+spring:
+  redis:
+    setting: maxmemory 128M
