+++ blog-api/src/main/resources/application.yaml
@@ -1,0 +1,14 @@
+spring:
+  redis:
+    host: 127.0.0.1
+    port: 6739
+    key:
+      keyword: keyword
+  profiles:
+    include:
+      - search
+
+jasypt:
+  encryptor:
+    password: doblee
+    bean: standardPBEStringEncryptor
