+++ blog-api/src/test/java/bank/common/config/JasyptConfigTest.java
@@ -21,4 +21,5 @@ class JasyptConfigTest {
         // then
         assertEquals(key, jasypt.decrypt(jasypt.encrypt(key)));
     }
+
 }
