--- blog-common/src/main/java/bank/common/mapper/GenericMapper.java
@@ -1,7 +1,0 @@
-package bank.common.mapper;
-
-public interface GenericMapper<F, T> {
-
-    T from(F from);
-
-}
