+++ blog-common/src/main/java/bank/mapper/GenericMapper.java
@@ -1,6 +1,5 @@
-package bank.remote.common;
+package bank.mapper;
 
-// TODO : 삭제
 public interface GenericMapper<F, T> {
 
     T from(F from);
