+++ blog-search/src/main/java/bank/config/FeignConfig.java
@@ -4,6 +4,6 @@ import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-@EnableFeignClients(basePackages = "bank")
+@EnableFeignClients(basePackages = "bank.remote")
 public class FeignConfig {
 }
