+++ blog-search/src/main/java/bank/domain/search/SortType.java
@@ -1,4 +1,4 @@
-package bank.common.type;
+package bank.domain.search;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
