+++ blog-search/src/main/java/bank/remote/kakao/KakaoClient.java
@@ -2,7 +2,6 @@ package bank.remote.kakao;
 
 import bank.remote.kakao.dto.KakaoSearchResponse;
 import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.stereotype.Component;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
