+++ blog-search/src/main/resources/application-search.yaml
@@ -1,13 +1,3 @@
-spring:
-  redis:
-    host: 127.0.0.1
-    port: 6739
-
-jasypt:
-  encryptor:
-    password: doblee
-    bean: standardPBEStringEncryptor
-
 feign:
   kakao:
     url: https://dapi.kakao.com
