+++ blog-search/src/test/java/bank/remote/kakao/KakaoSearchServiceTest.java
@@ -1,7 +1,7 @@
 package bank.remote.kakao;
 
-import bank.common.type.SortType;
 import bank.domain.search.SearchDocument;
+import bank.domain.search.SortType;
 import bank.remote.kakao.dto.KakaoDocument;
 import bank.remote.kakao.dto.KakaoSearchResponse;
 import bank.service.search.SearchCommand;
@@ -19,7 +19,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.when;
 
-@SpringBootTest
+@SpringBootTest(classes = KakaoSearchServiceImpl.class)
 class KakaoSearchServiceTest {
 
     @Autowired
